Refactor of module names, and organization folder
Removed CamelCase from module names, and put secondary modules in a new 'lib' folder
