+++ GEECS-Scanner-GUI/geecs_scanner/app/lib/module_open_folder.py
